+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed 🗑
+
+- Removed `UseMouseEventExtractorDefault`
+
 ## [0.11.4] - 2024-08-12
 
 ### New Features 🚀