+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `use_locale` now uses `unic_langid::LanguageIdentifier` and proper locale matching (thanks to @mondeja).
 - `use_preferred_dark` and `use_color_mode` now try to read the `Sec-CH-Prefers-Color-Scheme` header in SSR.
 
+### Fixes 🍕
+
+- Fixed the codec chapter in the book to refer to crate `codee`.
+
 ## [0.11.4] - 2024-08-12
 
 ### New Features 🚀